Refactor to move the map state out of the CollectAggregateFunction class
